refactor annotations and entrypoint
Rename annotation classes so they are simpler: Module, Object, Function.
As close as possible to the Dagger wording.

I also moved the info objects (the ones used to generate the json file)
to the dagger-java-sdk part so they can also be used to read the json
file.

Almost created the registry part of the entrypoint: it creates the right
dagger module structure, but now it needs to create the right JSON
output.
This entrypoint can be used as the entrypoint of a module. It's intended
to be generic (based on the json file for now, maybe on the annotation
in the future). So the module side only needs to declare the mainClass
to io.dagger.module.Entrypoint

For instance by adding this to the pom.xml:

      <dependencies>
        <dependency>
          <groupId>io.dagger</groupId>
          <artifactId>dagger-java-sdk</artifactId>
          <version>1.0.0-SNAPSHOT</version>
        </dependency>
        <dependency>
          <groupId>io.dagger</groupId>
          <artifactId>dagger-java-annotation-processor</artifactId>
          <version>1.0.0-SNAPSHOT</version>
          <scope>provided</scope>
        </dependency>
      </dependencies>
      <build>
        <pluginManagement>
          <plugin>
            <artifactId>maven-compiler-plugin</artifactId>
            <version>3.13.0</version>
            <configuration>
              <verbose>false</verbose>
              <annotationProcessorPaths>
                <path>
                  <groupId>io.dagger</groupId>
                  <artifactId>dagger-java-annotation-processor</artifactId>
                  <version>1.0.0-SNAPSHOT</version>
                </path>
              </annotationProcessorPaths>
              <annotationProcessors>
                <annotationProcessor>io.dagger.annotation.processor.DaggerModuleAnnotationProcessor</annotationProcessor>
              </annotationProcessors>
            </configuration>
          </plugin>
        </pluginManagement>
        <plugins>
          <plugin>
            <groupId>org.apache.maven.plugins</groupId>
            <artifactId>maven-shade-plugin</artifactId>
            <version>3.5.0</version>
            <executions>
              <execution>
                <phase>package</phase>
                <goals>
                  <goal>shade</goal>
                </goals>
                <configuration>
                  <transformers>
                    <transformer implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
                      <mainClass>io.dagger.module.Entrypoint</mainClass>
                    </transformer>
                  </transformers>
                </configuration>
              </execution>
            </executions>
          </plugin>
        </plugins>
      </build>

This should be good to create a jar with all the dependencies:

    ./mvnw clean package

This includes the management of the annotations.

And then the generated jar declares the right entrypoint, so this should
be good to run with:

    dagger run java -jar dagger-module.jar

This will not completely work as the module part is not done, so the
dagger engine doesn't know about the module and so the registration
can't be done.
But that's a different story. The goal here was to be able to:

 - document with annotations
 - create a first basic entrypoint
 - declare as much as possible in the SDK so it's easy to integrate

Signed-off-by: Yves Brissaud <gh@lgtd.io>