extract function invocation code
Do not invoke using reflection but directly by calling the function.
That's the benefit of the code generation instead of reflexivity.

This also allows to deal with param types better because there's no
need to find the right method signature which is easier in preparation
for parameterized types like Optional<String> and others.

Signed-off-by: Yves Brissaud <gh@lgtd.io>