feat(sdk/java): add @Internal annotation for fields
Fields can now be private. No need to have them public anymore.

All non transient/static fields will be serialized by default and
registered as a field to the Dagger API.
If you want to have a field serialized but not exposed it to the
API (so a user can retrieve the value) then it needs to be annotated with

    @Internal

Signed-off-by: Yves Brissaud <gh@lgtd.io>